apps/worker-python/requirements.txt
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.3.7
anthropic==0.7.8
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
